% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcs70}
\alias{bcs70}
\title{Data collected in the 1986 and 1996 waves of the BCS1970}
\format{
A named list of 7 data.tables. The unique common identifier is
\code{bcsid}. The tables are:
\describe{
  \item{bcs1986_arithmetic_data}{Results of an arithmetic test taken by all
    cohort members at 16}
  \item{bcs1986_occupational_interests}{Responses and results of a
    questionnaire on future occupations}
  \item{bcs1986_reading_matrices}{Results of a reading test taken by all
    cohort members at 16}
  \item{bcs1986derived}{Derived variables from the 1986 wave of the BCS1970}
  \item{bcs1986x}{Responses to the main survey of the 1986 wave}
  \item{bcs1996derived}{Derived variables from the 1996 wave of the BCS1970}
  \item{bcs1996x}{Responses to the main survey of the 1996 wave}
}
}
\source{
\url{https://beta.ukdataservice.ac.uk/datacatalogue/series/series?id=200001}
}
\usage{
bcs70
}
\description{
A list of 7 data.tables containing the information collected in the 1986 and
1996 waves of the BCS1970.
}
\keyword{datasets}
