% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcs70labels}
\alias{bcs70labels}
\title{Variable labels for \code{bcs70}}
\format{
A list of 7 data.tables.
\describe{
  \item{varName}{The name of the variable in \code{bcs70}}
  \item{varLabel}{The corresponding variable label containing additional
    information}
}
}
\source{
\url{https://beta.ukdataservice.ac.uk/datacatalogue/series/series?id=200001}
}
\usage{
bcs70labels
}
\description{
A list of 7 data.tables containing labels for the variables in \code{bcs70}.
}
\keyword{datasets}
