% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binnedDataExample}
\alias{binnedDataExample}
\title{Example data to demonstrate how to deal with binned outcome data}
\format{
A data.table of 2,614 rows and 2 variables:
\describe{
  \item{\code{id}}{A unique identifier for each observation}
  \item{\code{binnedData}}{Income data in bins}
}
}
\source{
See \code{bcs70}.
}
\usage{
binnedDataExample
}
\description{
A simplified dataset containing only binned income data (\code{binnedDate})
and an \code{id} variable to use in a demonstration of formatting binned data
to work with \code{progUkheEm}.
}
\keyword{datasets}
