% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtBcs4Em}
\alias{dtBcs4Em}
\title{BCS data formatted for use in \code{progUkheEm}.}
\format{
A data.table with 2,614 rows and 7 variables:
\describe{
  \item{bcsid}{The unique identifier for each member of the BCS1970 (chr)}
  \item{attSchl}{Reported attitude to school, z (factor, 5 levels)}
  \item{parInc}{Parental income in bins, y1 (factor, 13 levels)}
  \item{degree}{Whether the individual holds an UG degree at 25, d (logical)}
  \item{wklypay}{Weekly pay at 25 in GBP, y2 (numeric)}
  \item{left}{The lower bound for parental income in log(GBP), y1 (numeric)}
  \item{right}{The upper bound for parental income in log(GBP), y1 (numeric)}
}
}
\source{
See \code{bcs70}.
}
\usage{
dtBcs4Em
}
\description{
A data.table of cleaned and formatted BCS data ready for the EM algorithm,
as implemented by \code{progUkheEm}.
}
\keyword{datasets}
