% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtLsype4Em}
\alias{dtLsype4Em}
\title{LSYPE1 data formatted for use in \code{progUkheEm}}
\format{
A data.table of 1,584 rows and 18 variables:
\describe{
  \item{NSID}{Unique identifier for LSYPE cohort members (CMs)}
  \item{annualIncomeMP_w1}{Main parent's annual income (and partner) in wave
    1, banded (factor with 33 levels)}
  \item{annualIncomeMP_topBand_w1}{Further detail if in the top band of main
    parent's annual income, banded (factor with 60 levels)}
  \item{annualIncomeHH_w1}{Annual income of the household in wave 1
    (both parents?), banded (factor with 33 levels)}
  \item{contAnnIncHH_w1}{Annual income of the household in wave 1
    (both parents?), in GBP (numeric)}
  \item{annualIncomeHH_w4}{Annual income of the household in wave 4
    (both parents?), banded (factor with 12 levels)}
  \item{att2Debt16}{CMs attitude to debt at 16, score 2-24 (numeric)}
  \item{att2Schl16}{CMs attitude to school at 16, score 0-20 (numeric)}
  \item{infoStudFin16}{How well informed the CM felt about student finance
    at 16 (factor with 4 levels)}
  \item{sex}{CM's gender (factor with 2 levels)}
  \item{likeY11}{Whether CM enjoyed Year 11 (factor with 4 levels)}
  \item{mainAct16}{CM's main activity at 16 (factor with 5 levels)}
  \item{grssWkPay25}{CM's gross weekly pay at 25, in GBP (numeric)}
  \item{degree25}{Whether CM holds an UG degree at 25 (factor with 2 levels)}
  \item{mainAct25}{CM's main activity at 25 (factor with 14 levels)}
  \item{mainAct25_backcoded}{CM's main activity at 25 (factor with 10
    levels)}
  \item{y1}{Log of \code{contAnnIncHH_w1} (numeric)}
  \item{y2}{Log of \code{grssWkPay25} (numeric)}
}
}
\source{
See \code{lsype1YP} and \code{lsype1FB}.
}
\usage{
dtLsype4Em
}
\description{
A data.table of cleaned and formatted date ready for the EM algorithm as
implemented by \code{progUkheEm}.
}
\keyword{datasets}
