% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{ellInit}
\alias{ellInit}
\title{Initial values for parameters of Gaussian distribution}
\usage{
ellInit(x)
}
\arguments{
\item{x}{A data.table containing each binned observation as two values: a
column labelled \code{left} corresponding to the lower bound and
\code{right} corresponding to the upper.}
}
\description{
Calculates initial values for the Gaussian distribution from binnned data to
use with \code{\link[stats]{optim}} when performing ML.
}
