% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{kmeansSVs}
\alias{kmeansSVs}
\title{K-means Starting Values}
\usage{
kmeansSVs(y, K, y1cont = T)
}
\arguments{
\item{y}{A vector or matrix.}

\item{K}{A number (of types / groups).}

\item{y1cont}{A logical indicating the first component of y is continuous.}
}
\value{
A list of the starting values and other information.
  \itemize{
    \item \code{P_k} The proportions assigned to each group by
    \code{\link[stats]{kmeans}}
    \item \code{mu} The mean pre-treatment outcomes for each group
    \item \code{alpha} The mean post-treatment outcomes for each group
    \item \code{sigmaNu, sigmaEps} The standardard deviation of the pre- and
      post-treatment outcomes for each group
    \item \code{svTypes} The groups assigned to each observation by
      \code{\link[stats]{kmeans}}, i.e. \code{cluster}
    \item \code{muSigmaRes} An intermediate list containing results of ML to
      make y1 continuous if it is not. i.e. if \code{y1cont == FALSE}.
  }
}
\description{
Use \code{\link[stats]{kmeans}} to select starting values for the EM algorithm.
}
