% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsype1FB}
\alias{lsype1FB}
\title{Family background data from the LSYPE1 waves 1-8}
\format{
A list of 8 data.frames. The unique common identifier for each of the
15,770 individuals in wave 1 is \code{NSID}.
}
\source{
\url{https://beta.ukdataservice.ac.uk/datacatalogue/studies/study?id=5545}
}
\usage{
lsype1FB
}
\description{
A list containing 8 data.frames corresponding to the 8 waves of the LSYPE1.
}
\keyword{datasets}
