% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsype1YPlabels}
\alias{lsype1YPlabels}
\title{Variable labels for the LSYPE1 young person data}
\format{
A list of 8 data.tables:
\describe{
  \item{varNames}{The name of a variable in a data.frame in \code{lsype1YP}}
  \item{varLabel}{The corresponding variable label containing additional
    information}
}
}
\source{
\url{https://beta.ukdataservice.ac.uk/datacatalogue/studies/study?id=5545}
}
\usage{
lsype1YPlabels
}
\description{
A list of 8 data.tables containing variable names and labels corresponding to
the variables in \code{lsype1YP}.
}
\keyword{datasets}
