% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcs7}
\alias{mcs7}
\title{Data collected in the seventh wave of the Millennium Cohort Study}
\format{
A named list of 8 data.tables. The unique common identifier is
\code{MCSID}. The tables are:
\describe{
  \item{cm_cognitive_assessment}{Results of cognitive tests taken by cohort
    members (CM)}
  \item{cm_derived}{Derived variables from CM data}
  \item{cm_interview}{Variables from CM interview}
  \item{family_derived}{Derived variables from family interview}
  \item{family_interview}{Responses to the family interview}
  \item{hhgrid}{Data from the household grid}
  \item{parent_cm_interview}{Responses parent/CM joint interview}
  \item{parent_interview}{Responses parent interview}
}
}
\source{
\url{https://beta.ukdataservice.ac.uk/datacatalogue/series/series?id=2000031}
}
\usage{
mcs7
}
\description{
A list of 8 data.tables containing the information collected in the
}
\keyword{datasets}
