% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcs7Labels}
\alias{mcs7Labels}
\title{Variable labels for \code{mcs7}}
\format{
A list of 8 data.tables. Each contains the following two variables:
\describe{
  \item{varName}{The name of the variable in \code{mcs7}}
  \item{varLabel}{The corresponding variable label containing additional
    information}
}
}
\source{
\url{https://beta.ukdataservice.ac.uk/datacatalogue/series/series?id=2000031}
}
\usage{
mcs7Labels
}
\description{
A list of 8 data.tables containing labels for the variables in \code{mcs7}.
}
\keyword{datasets}
