% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progUkheEm.R
\name{progUkheEm}
\alias{progUkheEm}
\title{EM algorithm on wage and education data}
\usage{
progUkheEm(dt, K, startVals = "kmeans", maxiter = 400, y1cont = T)
}
\arguments{
\item{dt}{A data.table containing the wage and education data. The outcomes
columns containing the outcomes should be named \code{yp} for the before
treatment outcome, and \code{yw} for the post-treatment.}

\item{K}{An integer corresponding to the number of types.}

\item{startVals}{Either a character vector to use \code{kmeans()} or a vector
of starting values for the algorithm.}

\item{maxiter}{An integer specifying the maximum number of iterations before
stopping.}

\item{y1cont}{A logical value indicating whether the first outcome is
continuous. If not a special ML routine is used.}
}
\description{
This is the main function which takes data and other parameters and estimates
a mixture model including a treatment effect of schooling, on wage and
education data. The data needs to be in a specific format (see \code{dt})
and the model estimated is described in the accompanying vignettes.
}
