% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resBcs}
\alias{resBcs}
\title{The results of running \code{progUkheEm()} on BCS data}
\format{
A list of three elements:
\describe{
  \item{\code{listLogLike}}{A list of the log-likelihood at each iteration.}
  \item{\code{listParams}}{A list of the parameter values at each iteration.}
  \item{\code{dtLong}}{A data.table containing the data and final results of
  the EM algorithm.}
}
}
\source{
See \code{bcs70} and \code{progUkheEm.R}.
}
\usage{
resBcs
}
\description{
A list containing the results of running \code{progUkheEm()} on the BCS data
for K = 2, ..., 20. Test scores are the pre-treatment outcome. The elements
of the list correspond to that value of K, which represents the number of
groups. Each element is itself a list of three elements:
}
\keyword{datasets}
