% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukheEm-package.R
\docType{package}
\name{ukheEm-package}
\alias{ukheEm}
\alias{ukheEm-package}
\title{ukheEm: A package for estimating mixture models with exclusion restrictions
on education data}
\description{
The package includes a main function \code{progUkheEm} that estimates a
mixture model via the EM algorithm, and suitable data from two UK
longitudinal surveys.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/opmc/ukheEm}
}

}
\author{
\strong{Maintainer}: Oliver Cassagneau-Francis \email{ollie.francis@gmail.com}

}
\keyword{internal}
